ci(docker): improve handling of docker tags
Tag images in a better way by assigning an SHA hash for each build that
is made for a branch. This is safer as it would not produce a problem of
always relying upon the branch name of `main` and not knowing which
commit this `main` refers to.

Also the concrete versions of different steps in the pipeline are now
being used instead of using long SHA hashes.