peer: implement heartbeat handling for keepalive
Note that we start the keepalive routine once the peer is created. We do
it like this since the keepalive deadline is actually quite high and I
would say that if within that deadline no heartbeat messages were sent,
then we can consider the connection as stalled.

I.e. starting the keepalive timer only once the peer is connected is
like sparing a second that a peer normally needs to establish a
connection?