heartbeat: fix memory leak in ticker
`Ticker()` is a wrapper aroundn `NewTicker()` that is meant to be used
for infinite timers. This happens because `Ticker()` returns us a
receiving part of the channel, but no way to close this channel, so the
function leaks (which is also part of Go's documentation).

Our heartbeats must be finite: we must end them once the participant
leaves the conference, so `Ticker()` is not an option.