simulcast: change the way we treat simulcast
Instead of making it optional (with `nil`), we just create another value
in an enum that would signify the invalid value.

This is not perfect from the category theory point of view, but Go does
not allow ADTs and optional values for that. So we follow the
"convention" that is oftentimes used in such cases.

This would not work if we used more complex data structures, so
generally speaking it's a kludge to simplify handling things in code.