simulcast: treat layer requests smarter
So that if a user request layer X and X and lower layers are not
available, we'll send a higher layer. Otherwise the user may get
informed about track published, send a `subscribe` request, but the
information about published layer X has not yet been received by us, so
we would be forced to send nothing to the user when a layer is
requested.