Reproduceable builds with Nix
This adds a Nix flake containing reproduceable build expressions for Waterfall itself, a minimal Docker image, and a development environment. These can be accessed with nix build, nix build .#docker, and nix develop, respectively.

Admittedly I don't know much about Docker or running the SFU, so I didn't test it beyond seeing that it exited due to a missing config file when running Waterfall with docker run, but this is meant just as a PoC for now.