conference: optimize access to the subscriptions
This severely changes the way we treat subscriptions in order to speed
up things. Previous implementation dealt with publish and subscribe
logic in a simple and intuitive way, yet a very inefficient one for
bigger conferences.

Also, it stood on the way of an efficient simulcast implementation.
Since our main goal with simulcast is to make it scale better, this
change is sort of required in a long run although it does not really
bring a lot of performance improvements on top of the small conferences.

Note that in order to keep published and subscribed tracks, we keep them
together with the participants that actually file any changes to avoid
things going awry.