participant: remove canRequestKeyFrame logic
It seems like many key frame requests (i.e. when switching simulcast)
are quite critical and dropping them just because someone requested key
frame N milliseconds ago would result in us not delivering a key frame
that the participant may absolutely need to render the frame which would
result in participant seeing a grey frame for a while.

We probably can solve it in a more efficient way by e.g. storing the
key frames somewhere and firing them up (though they would be a bit too
old).

This is an experimental way to fix the issue.