track: decouple simulcast info from track info
This allows us to treat them separatel and make simulcast something that
changes over time without really changing the track info. As a side
effect, we can treat them separately and lift the restriction in our
code that required that the track's initial SSRC (after publishing)
matches the SSRC of the packet.

It turned out that the SSRC of the published tracks can change over
time, i.e. after the renegotiation.