peer: make key frame requests non-blocking
Peers use the sink to send the messages to the conference. Conference
listens to them and responds to them, which means that the conference
**must not** ever try to send to the channel (unless it's a non-blocking
sending) since it would otherwise hang the whole conference.

There is a single place where the conference however calls a function on
a `Peer` that in turn sends to the channel: when we request a key frame.

To prevent a potential deadlock, we only use a non-blocking call in this
case. All other usages of a channel happen inside a peer, i.e. always
in a different go-routine, so it's safe to use them.