peer: remove unnecessary accidental complexity
There was no point requesting the key frame by sending a message to the
conference as the only thing that the conference did was calling a
single function from the peer, so it was completely wasting the CPU
time. Instead, we could directly call this function and write to the
peer connection. The function is thread-safe since all mutable state is
protected by a mutex and the peer connection has a mutex inside.