conference: use worker for matrix messages
Otherwise we'll block waiting for the HTTP matrix layer to complete each
task when we try to send e.g. ICE candidates or doing something else.
The worker should normally never get full. The only reason it might get
full is either if we have **a lot** of participants or when the home
server is simply down.

We need a queue there (a buffered channel) since we don't need a
backpressure as such, we just want to offload CPU bound and I/O out of
the conference main loop.