conference: don't search for participant on msg
Don't check if the participant exists on each peer message that we
receive. In 99% of cases the participant is guaranteed to be in the map,
but in certain cases we don't even need a participant to process the
message (like for instance the RTP messages that we get very often).