track: implement new handling of published tracks
Now published track has its own package that encapsulated the logic
related to the published tracks and manages the lifetime of the track
and its subscriptions.

This means that from now on, each published track (i.e. each Pion's
`TrackRemote`) is handled separately from other tracks and by its own
go-routine, which means that handling of packets that belong to separate
tracks or even separate simulcast layers on a single track, are
processed in parallel.